+++ examples/multithreading.cpp
@@ -2,7 +2,7 @@
 #include <iostream>
 #include <thread>
 
-#include "../channel.hpp"
+#include "channel.hpp"
 
 int main()
 {
