+++ include/channel.cpp
